Update DraggableList as per review
int-11229
